# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AquaFarm Pro API
  version: 0.2.0
  description: |
    Core service contract for AquaFarm Pro.

    Included domains:
      - Authentication (JWT issuance & profile)
      - Multi‑tenancy (tenant creation / scoping)
      - Farms & Ponds (infrastructure inventory)
      - Water Quality (readings + trend analysis)
      - Fish Batches (stocking, growth lifecycle)
      - Feeding Records (nutrition tracking)
      - Accounting (journal entry draft capability)

    Notable design conventions:
      - Pagination envelope: { data: T[], meta } (all list endpoints unified; legacy array root keys deprecated and scheduled for removal in v0.4.0)
      - Standardized error responses via reusable components (ValidationError, UnauthorizedError, NotFoundError, ForbiddenError)
      - Consistent 201 bodies & 204 for destructive deletes
      - Examples supplied for major resource schemas & list responses

    Change log highlights:
      - 0.2.0: Added reusable list response components, standardized error $refs, multi-scenario water quality examples, enriched metadata.
      - 0.1.0: Initial rebuilt contract after corruption (auth, tenants, farms, ponds, core water quality & fish batches).
  termsOfService: https://example.com/terms
  contact:
    name: AquaFarm API Support
    url: https://example.com/support
    email: api-support@example.com
  license:
    name: UNLICENSED
    url: https://example.com/license
servers:
  - url: https://api.aquafarm.local/v1
...
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Farm:
      type: object
      description: Farm / production site.
      required: [id, name, ownerId, tenantId]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string, nullable: true }
        location: { type: string }
        totalArea: { type: number, format: float, nullable: true }
        farmType: { type: string, enum: [marine, freshwater, brackish] }
        status: { type: string, enum: [active, inactive, maintenance], default: active }
        coordinates:
          type: object
          nullable: true
          properties:
            latitude: { type: number, format: float }
            longitude: { type: number, format: float }
        contactPhone: { type: string, nullable: true }
        licenseNumber: { type: string, nullable: true }
        facilities:
          type: array
          items: { type: string }
        ownerId: { type: string, format: uuid }
        tenantId: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      example:
        id: "9a2c2ff3-3d24-45a6-93e1-0f42c4bf1001"
        name: Laguna Aquaculture Site
        description: Mixed species pilot site
        location: Laguna Province
        totalArea: 4.5
        farmType: freshwater
        status: active
        coordinates:
          latitude: 14.2101
          longitude: 121.3211
        contactPhone: "+63-900-111-2222"
        licenseNumber: FARM-2025-7788
        facilities: ["hatchery", "warehouse", "feed-storage"]
        ownerId: "b2b0a11e-1f44-4b55-9123-aa11ccff0001"
        tenantId: "92f7c9f0-1111-4c52-a5cd-9a11e3b90001"
        createdAt: "2025-09-01T09:00:00Z"
        updatedAt: "2025-09-15T07:50:00Z"
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        email: { type: string, format: email }
        role: { type: string }
        tenantId: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
      example:
        email: user@example.com
        password: StrongP@ssw0rd!
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
      example:
        access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refresh_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.refresh...
    ErrorResponse:
      type: object
      description: Standard error envelope returned for non-2xx responses.
      required: [error, message, code]
      properties:
        error:
          type: string
          description: High-level error name (e.g. ValidationError, NotFoundException)
        message:
          type: string
          description: Human-readable description (not guaranteed stable for programmatic parsing)
        statusCode:
          type: integer
          description: HTTP status code
        code:
          type: string
          description: Stable machine-readable application error code
          enum:
            - INTERNAL_ERROR
            - VALIDATION_ERROR
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - TENANT_NOT_FOUND
            - TENANT_INACTIVE
            - USER_NOT_FOUND
            - EMAIL_ALREADY_EXISTS
            - INVALID_CREDENTIALS
            - FARM_NOT_FOUND
            - POND_NOT_FOUND
            - WATER_READING_NOT_FOUND
            - FISH_BATCH_NOT_FOUND
            - FEEDING_RECORD_NOT_FOUND
            - ACCOUNT_NOT_FOUND
            - JOURNAL_ENTRY_NOT_FOUND
      example:
        error: ValidationError
        message: farmId must be a valid UUID
        statusCode: 400
        code: VALIDATION_ERROR
...
    FishBatch:
      type: object
      required: [id, batchNumber, species, initialCount, currentCount, averageWeight, totalBiomass, stockingDate, status, pondId]
      properties:
        id: { type: string, format: uuid }
        batchNumber: { type: string }
        species: { type: string }
        variety: { type: string, nullable: true }
        initialCount: { type: integer }
        currentCount: { type: integer }
        averageWeight: { type: number, format: float, description: grams }
        totalBiomass: { type: number, format: float, description: kilograms }
        stockingDate: { type: string, format: date-time }
        expectedHarvestDate: { type: string, format: date-time, nullable: true }
        actualHarvestDate: { type: string, format: date-time, nullable: true }
        status: { type: string, enum: [active, harvested, partial_harvest, dead] }
        survivalRate: { type: number, nullable: true }
        feedConversionRatio: { type: number, nullable: true }
        targetWeight: { type: number, nullable: true }
        supplier: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        healthStatus:
          type: object
          nullable: true
          properties:
            diseases: { type: array, items: { type: string } }
            treatments: { type: array, items: { type: string } }
            mortality: { type: number }
        pondId: { type: string, format: uuid }
        managedById: { type: string, format: uuid }
        tenantId: { type: string, format: uuid, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      example:
        id: "bf2ba7b5-6c9e-4c4b-b2a2-df02b0000001"
        batchNumber: BATCH-2025-0915-01
        species: Oreochromis niloticus
        variety: Red Tilapia
        initialCount: 7000
        currentCount: 6800
        averageWeight: 180
        totalBiomass: 1224
        stockingDate: "2025-06-10T09:00:00Z"
        expectedHarvestDate: "2025-12-01T09:00:00Z"
        actualHarvestDate: null
        status: active
        survivalRate: 0.97
        feedConversionRatio: 1.45
        targetWeight: 350
        supplier: HatcheryCo
        notes: Good growth trend
        healthStatus:
          diseases: []
          treatments: []
          mortality: 50
        pondId: "e6d0d8a2-5c2f-4ca2-8b6b-1d9f6b6e9a11"
        managedById: "c4f0b5c1-8ff1-43e2-9e09-4321a2f90001"
        tenantId: "92f7c9f0-1111-4c52-a5cd-9a11e3b90001"
        createdAt: "2025-06-10T09:00:00Z"
        updatedAt: "2025-09-15T07:20:00Z"
    CreateFishBatchRequest:
      type: object
      required: [batchNumber, species, initialCount, averageWeight, pondId]
      properties:
        batchNumber: { type: string }
        species: { type: string }
        variety: { type: string, nullable: true }
        initialCount: { type: integer }
        averageWeight: { type: number }
        pondId: { type: string, format: uuid }
        stockingDate: { type: string, format: date-time, nullable: true }
        expectedHarvestDate: { type: string, format: date-time, nullable: true }
    UpdateFishBatchRequest:
      type: object
      properties:
        status: { type: string }
        currentCount: { type: integer }
        averageWeight: { type: number }
        totalBiomass: { type: number }
        expectedHarvestDate: { type: string, format: date-time }
        actualHarvestDate: { type: string, format: date-time }
        targetWeight: { type: number }
        notes: { type: string }
...
  parameters:
...
    PageQuery:
      name: page
      in: query
      required: false
      description: |
        Page number (1-based). If the requested page exceeds the last page, the server clamps it to the last available page and returns that page's meta (no error raised).
      schema: { type: integer, minimum: 1, default: 1 }
...
    LimitQuery:
      name: limit
      in: query
      required: false
      description: |
        Page size (items per page). Values > 100 are capped to 100. Values < 1 are normalized to 1. Default varies by endpoint (commonly 25).
      schema: { type: integer, minimum: 1, maximum: 100 }
...
  responses:
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ForbiddenError:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    FishBatchListResponse:
      description: Paginated list of fish batches
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PaginatedFishBatchResponse'
          example:
            data:
              - id: "bf2ba7b5-6c9e-4c4b-b2a2-df02b0000001"
                batchNumber: BATCH-2025-0915-01
                species: Oreochromis niloticus
                initialCount: 7000
                currentCount: 6800
                averageWeight: 180
                totalBiomass: 1224
                stockingDate: "2025-06-10T09:00:00Z"
                status: active
                pondId: "e6d0d8a2-5c2f-4ca2-8b6b-1d9f6b6e9a11"
                createdAt: "2025-06-10T09:00:00Z"
                updatedAt: "2025-09-15T07:20:00Z"
            meta:
              total: 1
              page: 1
              limit: 10
              totalPages: 1
              hasNext: false
              hasPrev: false
...
    FarmListResponse:
      description: Paginated list of farms
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PaginatedFarmResponse'
          example:
            data:
              - id: "9a2c2ff3-3d24-45a6-93e1-0f42c4bf1001"
                name: Laguna Aquaculture Site
                location: Laguna Province
                farmType: freshwater
                status: active
                tenantId: "92f7c9f0-1111-4c52-a5cd-9a11e3b90001"
                ownerId: "b2b0a11e-1f44-4b55-9123-aa11ccff0001"
                createdAt: "2025-09-01T09:00:00Z"
                updatedAt: "2025-09-15T07:50:00Z"
            meta:
              total: 1
              page: 1
              limit: 10
              totalPages: 1
              hasNext: false
              hasPrev: false
...
paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: Login and obtain JWT pair
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: JWT issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid credentials / validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/register:
    post:
      tags: [Auth]
      summary: Register new user (assigns default tenant if not provided)
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '201':
          description: User registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
  /auth/profile:
    get:
      tags: [Auth]
      summary: Get current authenticated user profile
      operationId: getProfile
      responses:
        '200':
          description: Profile returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /farms:
    get:
      tags: [Farms]
      summary: List farms for current tenant (or all for admin with ?all=true)
      operationId: listFarms
      parameters:
        - name: all
          in: query
          required: false
          schema:
            type: string
          description: If 'true' and user role=admin returns all farms for tenant
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/LimitQuery'
      responses:
        '200': { $ref: '#/components/responses/FarmListResponse' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
    post:
      tags: [Farms]
      summary: Create farm (scoped to tenant / sets owner to current user)
      operationId: createFarm
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                location:
                  type: string
      responses:
        '201':
          description: Farm created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Farm'
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
  /farms/{id}:
    get:
      tags: [Farms]
      summary: Get farm by id (scoped by tenant)
      operationId: getFarm
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Farm details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Farm'
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
    patch:
      tags: [Farms]
      summary: Update farm (must own unless admin)
      operationId: updateFarm
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                location:
                  type: string
      responses:
        '200':
          description: Farm updated
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
    delete:
      tags: [Farms]
      summary: Delete farm (must own unless admin)
      operationId: deleteFarm
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Farm deleted
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
  /farms/{id}/stats:
    get:
      tags: [Farms]
      summary: Retrieve aggregate statistics for a farm
      operationId: getFarmStats
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Farm statistics returned
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
...
  /fish-batches:
    get:
      tags: [FishBatches]
      summary: List fish batches (paginated)
      operationId: listFishBatches
      parameters:
        - name: pondId
          in: query
          required: false
          schema: { type: string, format: uuid }
        - name: status
          in: query
          required: false
          schema: { type: string }
        - name: search
          in: query
          required: false
          schema: { type: string }
        - name: page
          in: query
          required: false
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          required: false
          schema: { type: integer, default: 10 }
      responses:
        '200': { $ref: '#/components/responses/FishBatchListResponse' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
    post:
      tags: [FishBatches]
      summary: Create fish batch
      operationId: createFishBatch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFishBatchRequest'
      responses:
        '201':
          description: Fish batch created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FishBatch'
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
  /fish-batches/{id}:
    get:
      tags: [FishBatches]
      summary: Get fish batch by id
      operationId: getFishBatch
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Fish batch details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FishBatch'
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
    patch:
      tags: [FishBatches]
      summary: Update fish batch
      operationId: updateFishBatch
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFishBatchRequest'
      responses:
        '200': { description: Fish batch updated }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
    delete:
      tags: [FishBatches]
      summary: Delete fish batch
      operationId: deleteFishBatch
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Fish batch deleted }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
  /feeding-records:
    get:
      tags: [Feeding]
      summary: List feeding records (paginated)
      operationId: listFeedingRecords
      parameters:
        - name: fishBatchId
          in: query
          required: false
          schema: { type: string, format: uuid }
        - name: page
          in: query
          required: false
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          required: false
          schema: { type: integer, default: 10 }
      responses:
        '200': { $ref: '#/components/responses/FeedingRecordListResponse' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
    post:
      tags: [Feeding]
      summary: Create feeding record
      operationId: createFeedingRecord
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFeedingRecordRequest'
      responses:
        '201':
          description: Feeding record created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedingRecord'
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }